+++ src/proxy/shield.go
@@ -392,6 +392,7 @@ func SmartShieldMiddleware(shieldID string, route utils.ProxyRouteConfig) func(h
 					"clientID": wrapper.ClientID,
 					"time": wrapper.TimeEnded.Sub(wrapper.TimeStarted).Seconds(),
 					"bytes": wrapper.Bytes,
+					"hostname": r.Host,
 					"url": r.URL,
 				})
 
@@ -444,6 +445,7 @@ func SmartShieldMiddleware(shieldID string, route utils.ProxyRouteConfig) func(h
 					"consumed": userConsumed,
 					"lastBan": lastBan,
 					"clientID": clientID,
+					"hostname": r.Host,
 					"url": r.URL,
 				})
 
@@ -510,6 +512,7 @@ func SmartShieldMiddleware(shieldID string, route utils.ProxyRouteConfig) func(h
 						"clientID": wrapper.ClientID,
 						"time": wrapper.TimeEnded.Sub(wrapper.TimeStarted).Seconds(),
 						"bytes": wrapper.Bytes,
+						"hostname": r.Host,
 						"url": r.URL,
 					})