+++ src/utils/middleware.go
@@ -246,12 +246,13 @@ func BlockByCountryMiddleware(blockedCountries []string, CountryBlacklistIsWhite
 
 							TriggerEvent(
 								"cosmos.proxy.shield.geo",
-								"Proxy Shield  Geo blocked",
+								"Proxy Shield Geo blocked",
 								"warning",
 								"",
 								map[string]interface{}{
 								"clientID": ip,
 								"country": countryCode,
+								"hostname": r.Host,
 								"url": r.URL.String(),
 							})
 
@@ -269,12 +270,13 @@ func BlockByCountryMiddleware(blockedCountries []string, CountryBlacklistIsWhite
 	
 							TriggerEvent(
 								"cosmos.proxy.shield.geo",
-								"Proxy Shield  Geo blocked",
+								"Proxy Shield Geo blocked",
 								"warning",
 								"",
 								map[string]interface{}{
 								"clientID": ip,
 								"country": countryCode,
+								"hostname": r.Host,
 								"url": r.URL.String(),
 							})
 
@@ -306,11 +308,12 @@ func BlockPostWithoutReferer(next http.Handler) http.Handler {
 				if ip != "" {
 					TriggerEvent(
 						"cosmos.proxy.shield.referer",
-						"Proxy Shield  Referer blocked",
+						"Proxy Shield Referer blocked",
 						"warning",
 						"",
 						map[string]interface{}{
 						"clientID": ip,
+						"hostname": r.Host,
 						"url": r.URL.String(),
 					})
 
@@ -538,6 +541,7 @@ func Restrictions(RestrictToConstellation bool, WhitelistInboundIPs []string) fu
 						"",
 						map[string]interface{}{
 						"clientID": ip,
+						"hostname": r.Host,
 						"url": r.URL.String(),
 					})
 
@@ -556,6 +560,7 @@ func Restrictions(RestrictToConstellation bool, WhitelistInboundIPs []string) fu
 						"",
 						map[string]interface{}{
 						"clientID": ip,
+						"hostname": r.Host,
 						"url": r.URL.String(),
 					})
 
@@ -576,6 +581,7 @@ func Restrictions(RestrictToConstellation bool, WhitelistInboundIPs []string) fu
 				"",
 				map[string]interface{}{
 				"clientID": ip,
+				"hostname": r.Host,
 				"url": r.URL.String(),
 			})