+++ vite.config.js
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import visualizer from 'rollup-plugin-visualizer';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -7,18 +8,21 @@ export default defineConfig({
   root: 'client',
   build: {
     outDir: '../static',
+    rollupOptions: {
+      plugins: [visualizer({ open: true })],
+    },
   },
   server: {
     proxy: {
       '/cosmos/api': {
-        target: 'https://localhost:8443',
-        // target: 'https://192.168.1.170:8443',
+        // target: 'https://localhost:8443',
+        target: 'http://192.168.1.170:8080',
         secure: false,
         ws: true,
       },
       '/cosmos/rclone': {
-        target: 'https://localhost:8443',
-        // target: 'https://192.168.1.170:8443',
+        // target: 'https://localhost:8443',
+        target: 'http://192.168.1.170:8080',
         secure: false,
         ws: true,
       }